feat: add new zksync-llvm-runner with llvm18 support (#84)
## Changes

* [x] Introduce [new `zksync-llvm-runner`
](https://github.com/matter-labs/compiler-infra/pkgs/container/zksync-llvm-runner)
package with updated CI
* [x] Introduce semver support with tags keeping the `latest` available
* [x] Deprecate CI for outdated images
* [x] Update host LLVM from 17 to 18 to be compatible with Rustc 1.79
we're using

## Purpose

Images will continue to support `latest` notion, but at the same time we
introduce semver for them to be able to properly test and switch to the
new versions without breaking any workflows. When the new image is
released, it will be tagged as `main` allowing to perform testing in CI
repository. After the testing is successful, we can tag it and `latest`
will be automatically updated distributing the changes without the need
of updating any CI workflows.

At the same time, we will update host LLVM to be compatible with the
LLVM we're using in Rust to solve all kinds of the binary conflicts when
re-using some LLVM features between C++ and Rust like code coverage
instrumentation.